+++ src/scanner/monga_tokens.h
@@ -76,7 +76,7 @@ union YYSTYPE
 {
 #line 17 "src/parser/monga.y" /* yacc.c:1909  */
 
-  double              float_val;
+  float               float_val;
   int                 int_val;
   char               *string_val;
   char                char_val;
